+++ src/test/kotlin/de/magynhard/crystal/parser/CrystalParserTest.kt
@@ -712,6 +712,10 @@ class CrystalParserTest : ParsingTestCase("", "cr", CrystalParserDefinition()) {
         doTest(true)
     }
 
+    fun testBacktickInInterpolation() {
+        doTest(true)
+    }
+
     fun testMacroExpressions() {
         doTest(true)
     }